[server] Handle db cleanup failures
At the end of the run remove API we do some database clean up. It is possible that
it fails if we set a statement timeout for our postgresql database and the
clean up query takes too much time time to execute, so the database will
abort the transaction and throws an exception. To solve this problem we will
handle failures for database cleanup queries so it doesn't effect other API
requests.
