+++ web/server/codechecker_server/database/database.py
@@ -56,6 +56,28 @@ def call_command(cmd, env=None, cwd=None):
         return oerr.strerror, oerr.errno
 
 
+class DBSession(object):
+    """
+    Requires a session maker object and creates one session which can be used
+    in the context.
+
+    The session will be automatically closed, but commiting must be done
+    inside the context.
+    """
+    def __init__(self, session_maker):
+        self.__session = None
+        self.__session_maker = session_maker
+
+    def __enter__(self):
+        # create new session
+        self.__session = self.__session_maker()
+        return self.__session
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        if self.__session:
+            self.__session.close()
+
+
 class DBContext(object):
     """
     Simple helper class to setup and sql engine, a database session
@@ -613,3 +635,10 @@ def conv(filter_value):
     if filter_value is None:
         return '%'
     return filter_value.replace('*', '%')
+
+
+def escape_like(string, escape_char='*'):
+    """Escape the string parameter used in SQL LIKE expressions."""
+    return string.replace(escape_char, escape_char * 2) \
+                 .replace('%', escape_char + '%') \
+                 .replace('_', escape_char + '_')
